🧪 [STORYBOOK] rename files
