+++ src/stories/FirstButton.stories.jsx
@@ -1,0 +1,5 @@
+export default {
+    title: "Example/Test",
+};
+
+export const TestStory = () => <button>Test Button</button>;
