+++ src/stories/TestButton.stories.jsx
@@ -1,5 +1,11 @@
+import Button from "../components/ui/test/Button";
+
 export default {
-    title: "Example/Test",
+  title: "Example/Button",
+  component: Button,
 };
 
-export const TestStory = () => <button>Test Button</button>;
+export const Small = () => <Button size="sm">Small Button</Button>;
+export const Medium = () => <Button size="md">Medium Button</Button>;
+export const Large = () => <Button size="lg">Large Button</Button>;
+export const ExtraLarge = () => <Button size="xl">Extra Large Button</Button>;
